move each module into subdirectory under src
