+++ src/Marian/types.ts
@@ -1,0 +1,7 @@
+import { RefreshInfo } from '../SearchIndex/types';
+// import { RefreshInfo } from "../SearchIndex/types";
+
+export interface StatusResponse {
+  manifests: string[];
+  lastSync?: RefreshInfo | null;
+}
