+++ src/SearchIndex/index.ts
@@ -1,230 +1,15 @@
 import assert from 'assert';
-import crypto from 'crypto';
-import fs from 'fs';
-import util from 'util';
-import S3 from 'aws-sdk/clients/s3';
-import { MongoClient, Collection, TransactionOptions, AnyBulkWriteOperation, Db, ClientSession } from 'mongodb';
-// @ts-ignore
-import dive from 'dive';
 // @ts-ignore
 import Logger from 'basic-logger';
-import { Query } from './Query';
-import { convertTaxonomyResponse, formatFacetMetaResponse } from './util';
+import { MongoClient, Collection, TransactionOptions, AnyBulkWriteOperation, Db, ClientSession } from 'mongodb';
+
+import { convertTaxonomyResponse, formatFacetMetaResponse, getManifests, joinUrl } from './util';
+import { Query } from '../Query';
+import { Document, Manifest, DatabaseDocument, RefreshInfo, Taxonomy, FacetDisplayNames, FacetAggRes } from './types';
 
 const log = new Logger({
   showTimestamp: true,
 });
-
-export interface Document {
-  slug: string;
-  title?: string;
-  headings?: string[];
-  text?: string; // legacy manifests have text instead of paragraphs
-  paragraphs?: string; // legacy manifests will not have paragraphs
-  code?: {}[]; // legacy manifests will not have code field
-  preview?: string;
-  tags: null | string[];
-  facets?: {};
-}
-
-interface ManifestData {
-  documents: Document[];
-  includeInGlobalSearch: boolean;
-  url: string;
-}
-
-interface Manifest {
-  manifest: ManifestData;
-  lastModified: Date;
-  manifestRevisionId: string;
-  searchProperty: string;
-}
-
-export interface DatabaseDocument extends Document {
-  url: string;
-  manifestRevisionId: string;
-  searchProperty: string[];
-  includeInGlobalSearch: boolean;
-}
-
-export interface RefreshInfo {
-  deleted: number;
-  updated: string[];
-  skipped: string[];
-  errors: Error[];
-  dateStarted: Date;
-  dateFinished: Date | null;
-  elapsedMS: number | null;
-}
-
-export interface TaxonomyEntity {
-  name: string;
-  display_name?: string;
-  [x: string]: TaxonomyEntity[] | string | undefined;
-}
-export type Taxonomy = Record<string, TaxonomyEntity[]>;
-
-export type FacetDisplayNames = {
-  name?: string;
-  [key: string]: FacetDisplayNames | string | boolean | undefined;
-};
-
-export type FacetBucket = {
-  buckets: {
-    _id: string;
-    count: number;
-  }[];
-};
-export type FacetAggRes = {
-  count: {
-    lowerBound: number;
-  };
-  facet: {
-    [key: string]: FacetBucket;
-    // key is split by '>' key
-  };
-};
-
-export function joinUrl(base: string, path: string): string {
-  return base.replace(/\/*$/, '/') + path.replace(/^\/*/, '');
-}
-
-function generateHash(data: string): Promise<string> {
-  const hash = crypto.createHash('sha256');
-
-  return new Promise((resolve, reject) => {
-    hash.on('readable', () => {
-      const data = hash.read();
-      if (data) {
-        resolve(data.toString('hex'));
-      }
-    });
-
-    hash.write(data);
-    hash.end();
-  });
-}
-
-async function getManifestsFromS3(bucketName: string, prefix: string): Promise<Manifest[]> {
-  const s3 = new S3({ apiVersion: '2006-03-01' });
-  const result: S3.Types.ListObjectsV2Output = await util.promisify(
-    s3.makeUnauthenticatedRequest.bind(s3, 'listObjectsV2', {
-      Bucket: bucketName,
-      Prefix: prefix.replace(/^\//, ''),
-    })
-  )();
-
-  if (result.IsTruncated) {
-    // This would indicate something awry, since we shouldn't
-    // ever have more than 1000 properties. And if we ever did,
-    // everything would need to be rearchitected.
-    throw new Error('Got truncated response from S3');
-  }
-
-  const manifests = [];
-  for (const bucketEntry of result.Contents || []) {
-    if (bucketEntry.Size === 0) {
-      log.error(new Error(`Got empty file: "${bucketEntry.Key}"`));
-      continue;
-    }
-
-    assert.ok(bucketEntry.Key);
-
-    const matches = bucketEntry.Key.match(/([^/]+).json$/);
-    if (matches === null) {
-      log.error(new Error(`Got weird filename in manifest listing: "${bucketEntry.Key}"`));
-      continue;
-    }
-
-    const searchProperty = matches[1];
-    const data: S3.Types.GetObjectOutput = await util.promisify(
-      s3.makeUnauthenticatedRequest.bind(s3, 'getObject', {
-        Bucket: bucketName,
-        Key: bucketEntry.Key,
-      })
-    )();
-
-    assert.ok(data.Body);
-    assert.ok(data.LastModified);
-
-    const body = data.Body.toString('utf-8');
-    const hash = await generateHash(body);
-    const parsed = JSON.parse(body);
-    manifests.push({
-      manifest: parsed,
-      lastModified: data.LastModified,
-      manifestRevisionId: hash,
-      searchProperty: searchProperty,
-    });
-  }
-
-  return manifests;
-}
-
-function getManifestsFromDirectory(prefix: string): Promise<Manifest[]> {
-  return new Promise((resolve, reject) => {
-    const manifests: Manifest[] = [];
-
-    dive(
-      prefix,
-      async (err: Error | null, path: string, stats: fs.Stats) => {
-        if (err) {
-          reject(err);
-        }
-        const matches = path.match(/([^/]+).json$/);
-        if (!matches) {
-          return;
-        }
-        const searchProperty = matches[1];
-        const data = fs.readFileSync(path, { encoding: 'utf-8' });
-        const parsed = JSON.parse(data);
-        const hash = await generateHash(data);
-
-        manifests.push({
-          manifest: parsed,
-          lastModified: stats.mtime,
-          manifestRevisionId: hash,
-          searchProperty: searchProperty,
-        });
-      },
-      () => {
-        resolve(manifests);
-      }
-    );
-  });
-}
-
-/// Fetch manifests from a given path. It can (for historic cruft reasons)
-/// take one of two formats:
-/// dir:<path> to load manifests from a local directory.
-/// s3://<bucketName>/<prefix> to load manifests from an S3 location.
-async function getManifests(manifestSource: string): Promise<Manifest[]> {
-  log.info(`Loading manifests from ${manifestSource}`);
-  const parsed = new URL(manifestSource);
-
-  let manifests;
-  if (parsed.protocol === 's3:') {
-    const bucketName = parsed.host.trim();
-    const prefix = parsed.pathname.trim();
-    if (!bucketName.length || !prefix.length) {
-      throw new Error('Bad bucket manifest source');
-    }
-    manifests = await getManifestsFromS3(bucketName, prefix);
-  } else if (parsed.protocol === 'dir:') {
-    manifests = await getManifestsFromDirectory(parsed.pathname);
-  } else {
-    throw new Error('Unknown manifest source protocol');
-  }
-
-  // We have a persistent problem with weird URLs. Remove excess leading slashes.
-  for (const manifest of manifests) {
-    const urlRoot = new URL(manifest.manifest.url);
-    urlRoot.pathname = urlRoot.pathname.replace(/^\/+/, '');
-    manifest.manifest.url = urlRoot.toString();
-  }
-
-  return manifests;
-}
 
 export class SearchIndex {
   currentlyIndexing: boolean;
